eef1c77 Removed debug line.
3f85c0d Exclude warnings from documentation generation tool.
e498680 Exclude generated files/protobuf from coverage.
57abd70 add more warnings to exclusions list
3efb835 Fixed warnings
c168d1c Fixed warnings
2622357 Fixed warnings
b4086a1 Fixed warnings
afecee8 Fixed warnings
4f2ec04 Adding more exceptions.
1ecdede Trying to fix dashboard exclusions.
8178d76 Merge remote branch 'stage/fix-warnings' into dashboard_exclusions
00cf697 Fix conversion warning in vtkSICompoundSourceProxy
d724eca Fix unused variable warning in vtkSpyPlotReader
540d5c6 Disable assignment operator for vtkSpyPlotIStream
6562aed Fix conversion warning in vtkHierarchicalFractal
...

+++ ParaViewCore/ServerManager/Core/vtkSMSessionProxyManager.cxx
@@ -364,7 +364,7 @@ unsigned int vtkSMSessionProxyManager::GetNumberOfProxies(const char* group)
       {
       size += it2->second.size();
       }
-    return size;
+    return static_cast<unsigned int>(size);
     }
   return 0;
 }
@@ -901,7 +901,7 @@ void vtkSMSessionProxyManager::UpdateProxyInOrder(vtkSMProxy* proxy)
 //---------------------------------------------------------------------------
 int vtkSMSessionProxyManager::GetNumberOfLinks()
 {
-  return this->Internals->RegisteredLinkMap.size();
+  return static_cast<int>(this->Internals->RegisteredLinkMap.size());
 }
 
 //---------------------------------------------------------------------------
